requirements.txt
pbr>=1.10.0,<=5.8.1  # Apache-2.0
oslo.config>=3.22.2,<10.0.0  # Apache-2.0
restalchemy>=12.9.0,<13.0.0  # Apache-2.0
gcl_iam>=0.8.0,<1.0.0  # Apache-2.0
gcl_looper>=0.1.0,<=1.0.0  # Apache-2.0
bjoern>=3.2.2  # BSD License (BSD-3-Clause)
Jinja2>=3.1.5,<4.0.0  # BSD License (BSD-3-Clause)
bazooka>=1.3.0,<2.0.0  # Apache-2.0
pyOpenSSL>=25.0.0,<26.0.0  # Apache-2.0
rich>=14.0.0,<15.0.0  # MIT License (MIT)
